image_processor/requirements.txt
rembg
pillow
pillow-heif
onnxruntime
